requirements.txt
tensorflow>=2.11.0
Pillow>=9.1.1
pytz>=2022.6
tqdm>=4.64.1
scikit-image>=0.19.3
tf-models-official>=2.11.3